+++ packages/patternfly-4/react-docs/gatsby-config.js
@@ -43,8 +43,8 @@ module.exports = {
         ignore: ignore
       },
     },
+    // Our custom plugin for *.js?x *.ts?x files to get prop types
     {
-      // Our custom plugin for *.js?x *.ts?x files to get prop types
       resolve: path.resolve(`${__dirname}/plugins/gatsby-transformer-react-docgen-typescript`),
     },
     // The markdown plugin for *.md files