+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/LabelDemo/LabelDemo.tsx
@@ -1,0 +1,19 @@
+import { Label, LabelProps } from '@patternfly/react-core';
+import React, { Component } from 'react';
+export class LabelDemo extends Component {
+ myLabelProps: LabelProps = {
+   isCompact: true,
+   children: "Compact label"
+ }
+
+ render() {
+   return (
+     <React.Fragment>
+       <Label>Default label</Label>
+       <Label isCompact={this.myLabelProps.isCompact}>{this.myLabelProps.children}</Label>
+     </React.Fragment>
+   );
+ }
+}
+
+export default LabelDemo;