+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/index.ts
@@ -1,5 +1,6 @@
 export * from './AlertDemo/AlertDemo';
 export * from './AvatarDemo/AvatarDemo';
+export * from './LabelDemo/LabelDemo';
 export * from './NavDemo/NavDemo';
 export * from './PopoverDemo/PopoverDemo'
 export * from './TabsDemo/TabsDemo';