+++ packages/patternfly-4/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.34](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.33...@patternfly/react-inline-edit-extension@2.3.34) (2019-05-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.3.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.32...@patternfly/react-inline-edit-extension@2.3.33) (2019-05-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.3.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.31...@patternfly/react-inline-edit-extension@2.3.32) (2019-05-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.3.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.30...@patternfly/react-inline-edit-extension@2.3.31) (2019-05-09)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.3.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.29...@patternfly/react-inline-edit-extension@2.3.30) (2019-05-08)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [2.3.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.3.28...@patternfly/react-inline-edit-extension@2.3.29) (2019-05-08)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension