+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.1.1...demo-app-ts@1.1.2) (2019-05-09)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.1.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.1.0...demo-app-ts@1.1.1) (2019-05-09)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [1.1.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.0.3...demo-app-ts@1.1.0) (2019-05-09)
+
+
+### Features
+
+* **Label:** Convert Label to TypeScript ([#1896](https://github.com/patternfly/patternfly-react/issues/1896)) ([a950fd5](https://github.com/patternfly/patternfly-react/commit/a950fd5)), closes [#1893](https://github.com/patternfly/patternfly-react/issues/1893)
+
+
+
+
+
+## 1.0.3 (2019-05-09)
+
+
+### Bug Fixes
+
+* **docs:** fix ts docgen ([#1944](https://github.com/patternfly/patternfly-react/issues/1944)) ([f68bbc5](https://github.com/patternfly/patternfly-react/commit/f68bbc5))
+
+
+
+
+
+## [1.0.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/demo-app-ts@1.0.1...@patternfly/demo-app-ts@1.0.2) (2019-05-08)
+
+**Note:** Version bump only for package @patternfly/demo-app-ts
+
+
+
+
+
 ## 1.0.1 (2019-05-08)